parallel insert and delete

+++ interactive_engine/executor/store/bmcsr/src/bmcsr.rs
@@ -34,7 +34,7 @@ impl<I: IndexType> BatchMutableCsrBuilder<I> {
         }
     }
 
-    pub fn init(&mut self, degree: &Vec<i64>, _: f64) {
+    pub fn init(&mut self, degree: &Vec<i32>, _: f64) {
         let vertex_num = degree.len();
         let mut edge_num = 0_usize;
         for i in 0..vertex_num {
